Finalização da integracao do com o backend
